feature/#60: backStackEntry 값 전달

+++ client/app/src/main/java/com/example/circuler/presentation/ui/request/navigation/RequestNavigation.kt
@@ -17,7 +17,7 @@ fun NavController.navigateToRequestPackage(navOptions: NavOptions? = null) {
 
 fun NavGraphBuilder.requestGraph(
     navigateUp: () -> Unit,
-    navigateToEnterPackaging: () -> Unit
+    navigateToEnterPackaging: (Int) -> Unit
 ) {
     composable<Route.EntirePackageRequest> {
         RequestRoute(
